broke usermenu into components
