+++ src/components/Navigation/Navigation.tsx
@@ -1,38 +1,8 @@
-import {
-  AppBar,
-  Toolbar,
-  Button,
-  Menu,
-  MenuItem,
-  Link as MUILink,
-} from "@mui/material";
-import PersonIcon from "@mui/icons-material/Person";
-import { useState } from "react";
-import { switchToRegistration } from "@/reducers/registrationReduser";
-import { removeUserInfo } from "@/reducers/loginReducer";
-import { connect } from "react-redux";
-import { switchToRegistrationType } from "@/pages/login/Login.types";
+import { AppBar, Toolbar, Link as MUILink } from "@mui/material";
 import Link from "next/link";
-import { ILoginSlice } from "@/reducers/loginReducer.types";
+import UserMenu from "./UserMenu/UserMenu";
 
-const Navigation = ({
-  login,
-  removeUserInfo,
-  switchToRegistration,
-}: {
-  login: ILoginSlice;
-  removeUserInfo: switchToRegistrationType;
-  switchToRegistration: switchToRegistrationType;
-}) => {
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
+const Navigation = () => {
   return (
     <AppBar position="static">
       <Toolbar variant="dense">
@@ -47,63 +17,10 @@ const Navigation = ({
         >
           Lanle WEB
         </MUILink>
-        {login.state ? `${login.state}` : null}
-        <Button
-          color="inherit"
-          id="basic-button"
-          aria-controls={open ? "basic-menu" : undefined}
-          aria-haspopup="true"
-          aria-expanded={open ? "true" : undefined}
-          onClick={handleClick}
-        >
-          <PersonIcon />
-        </Button>
-        {login.state ? (
-          <Menu
-            id="basic-menu"
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-            MenuListProps={{
-              "aria-labelledby": "basic-button",
-            }}
-          >
-            <MenuItem onClick={() => removeUserInfo(true)}>
-              <Link href="/login">Logout</Link>
-            </MenuItem>
-          </Menu>
-        ) : (
-          <Menu
-            id="basic-menu"
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-            MenuListProps={{
-              "aria-labelledby": "basic-button",
-            }}
-          >
-            <MenuItem onClick={() => switchToRegistration(false)}>
-              <Link href="/login">Login</Link>
-            </MenuItem>
-            <MenuItem onClick={() => switchToRegistration(true)}>
-              <Link href="/login">Sign In</Link>
-            </MenuItem>
-          </Menu>
-        )}
+        <UserMenu />
       </Toolbar>
     </AppBar>
   );
 };
 
-const mapStateToProps = ({ login }: { login: ILoginSlice }) => {
-  return {
-    login: login,
-  };
-};
-
-const mapDispatchProps = {
-  removeUserInfo,
-  switchToRegistration,
-};
-
-export default connect(mapStateToProps, mapDispatchProps)(Navigation);
+export default Navigation;
